requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
attrs==25.4.0
beautifulsoup4==4.14.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
contourpy==1.3.3
curl_cffi==0.13.0
cycler==0.12.1
dateparser==1.2.2
eventkit==1.0.3
fonttools==4.60.1
frozendict==2.4.6
frozenlist==1.8.0
ib-insync==0.9.86
idna==3.11
kiwisolver==1.4.9
matplotlib==3.10.7
multidict==6.7.0
multitasking==0.0.12
nest-asyncio==1.6.0
numpy==2.3.4
packaging==25.0
pandas==2.3.3
peewee==3.18.2
pillow==12.0.0
platformdirs==4.5.0
propcache==0.4.1
protobuf==6.33.0
pycparser==2.23
pycryptodome==3.23.0
pyparsing==3.2.5
python-binance==1.0.30
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
regex==2025.10.23
requests==2.32.5
six==1.17.0
soupsieve==2.8
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
websockets==15.0.1
yarl==1.22.0
yfinance==0.2.66
pymongo==4.5.0
motor==3.3.2
fastapi==0.111.1
uvicorn==0.30.1
apscheduler==3.10.4
tiktoken==0.12.0
pytest==9.0.2
pytest-asyncio==1.3.0